reconstructed Autotools build system (closes #1)
The resulting config.h.in is similar to the one from the CVS repo.
However, I removed checks that apparently weren't referenced in the code base.
Also, I fixed the terminfo check. Previously, it was actually building against
termcap on my system. Terminfo should always be preferred.
Currently, it does not check against ncurses itself. If it turns out there are
systems with an ncurses pkg-config file, but without tinfo, we might have to check
for ncurses as well.

It adds both -Wall and -Wsign-compare, which turned out to be useful to
debug the DOS version.